pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentkit"
version = "0.1.0"
description = "Deterministic orchestration engine for AI-driven story execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Stefan Altenhof" },
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "ruff>=0.4",
    "pre-commit>=3.7",
]

[project.scripts]
agentkit = "agentkit.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "requires_git: marks tests requiring git",
    "requires_gh: marks tests requiring GitHub CLI",
    "integration: integration tests",
    "contract: contract/snapshot tests",
    "e2e: end-to-end tests against live systems",
]

[tool.coverage.run]
source = ["agentkit"]

[tool.coverage.report]
fail_under = 85
show_missing = true

[tool.mypy]
strict = true
packages = ["agentkit"]
mypy_path = "src"

[tool.ruff]
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]
